Moved speed change class to MultiplierShiftSetting type
